fixed doc comment errors, seperated files into multiple folders
